tutorial/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.13
djangorestframework==3.13.1
Pygments==2.12.0
pytz==2022.1
sqlparse==0.4.2
